mock data as fixture - party walls
