+++ packages/party_walls/tests/conftest.py
@@ -1,10 +1,11 @@
 import os
+import pickle
 from pathlib import Path
 
 import pytest
 from bag3d.common.resources.database import DatabaseConnection
 from bag3d.common.resources.files import file_store
-from dagster import build_op_context
+from dagster import AssetKey, IOManager, SourceAsset, build_op_context
 
 LOCAL_DIR = os.getenv("BAG3D_TEST_DATA")
 HOST = "localhost"
@@ -91,3 +92,57 @@ def pytest_collection_modifyitems(config, items):
         for item in items:
             if "slow" in item.keywords:
                 item.add_marker(skip_slow)
+
+
+@pytest.fixture(scope="session")
+def mock_features_file_index(intermediate_data_dir, input_data_dir):
+    class MockIOManager(IOManager):
+        def load_input(self, context):
+            data = pickle.load(
+                open(intermediate_data_dir / "features_file_index.pkl", "rb")
+            )
+            for k, v in data.items():
+                data[k] = Path(str(v).replace(str(v.parents[8]), str(input_data_dir)))
+            return data
+
+        def handle_output(self, context, obj):  # pragma: no cover
+            raise NotImplementedError()
+
+    return SourceAsset(
+        key=AssetKey(["party_walls", "features_file_index"]),
+        io_manager_def=MockIOManager(),
+    )
+
+
+@pytest.fixture(scope="session")
+def mock_distribution_tiles_files_index(intermediate_data_dir, input_data_dir):
+    class MockIOManager(IOManager):
+        def load_input(self, context):
+            data = pickle.load(
+                open(intermediate_data_dir / "distribution_tiles_files_index.pkl", "rb")
+            )
+            for i, d in enumerate(data.paths_array):
+                data.paths_array[i] = Path(
+                    str(d).replace(str(d.parents[7]), str(input_data_dir))
+                )
+            for k, v in data.export_results.items():
+                cj_path = data.export_results[k].cityjson_path
+                data.export_results[k].cityjson_path = Path(
+                    str(cj_path).replace(str(cj_path.parents[7]), str(input_data_dir))
+                )
+                gpkg_path = data.export_results[k].gpkg_path
+                data.export_results[k].gpkg_path = Path(
+                    str(gpkg_path).replace(
+                        str(gpkg_path.parents[7]), str(input_data_dir)
+                    )
+                )
+                # TODO: fix data.export_results[k].obj_paths
+            return data
+
+        def handle_output(self, context, obj):  # pragma: no cover
+            raise NotImplementedError()
+
+    return SourceAsset(
+        key=AssetKey(["party_walls", "distribution_tiles_files_index"]),
+        io_manager_def=MockIOManager(),
+    )
